feat(model): add domain models and switch price to kopecks
- Add internal/model/ with 7 files covering all 12 DB tables
- user.go: User (json:- on password_hash), AuditLog
- direction.go: Direction
- doctor.go: Doctor, DoctorDirection
- service.go: Service — price *int64 (kopecks, not float64)
- schedule.go: WorkingHours, ScheduleException + ExceptionType enum
- patient.go: Patient, BotSession
- appointment.go: Appointment, AppointmentStatusHistory + status/source enums
- Change migrations/init: services.price DECIMAL(10,2) -> BIGINT (kopecks)
- Update docs/database.md: note price as BIGINT (kopecks)

+++ backend/migrations/20260519100000_init.sql
@@ -58,7 +58,7 @@ CREATE TABLE services (
     name             VARCHAR(255) NOT NULL,
     description      TEXT,
     duration_minutes INT NOT NULL CHECK (duration_minutes > 0),
-    price            DECIMAL(10, 2),
+    price            BIGINT,
     is_active        BOOLEAN NOT NULL DEFAULT true,
     created_at       TIMESTAMPTZ NOT NULL DEFAULT NOW(),
     updated_at       TIMESTAMPTZ NOT NULL DEFAULT NOW()
